Move emoji letter rendering to new wrapped connector component
